requirements.txt
streamlit
pandas
matplotlib
plotly
scikit-learn
pillow
